db.App is removed in favor of stupid.App stupid.App has their own extention function Extention functions are now actually implemented LastCon log value is now properly the date. Added AllowedDomains to apiKeys

+++ user.go
@@ -3,7 +3,7 @@ package stupid
 type logUser struct {
 	ID       string `json:"id" bson:"_id"`
 	Platform string `json:"platform" bson:"platform"`
-	LastCon  int64  `json:"lastCon" bson:"lastCon"`
+	LastCon  int    `json:"lastCon" bson:"lastCon"`
 }
 
 type AuthdUser struct {
